fix: address elicitation review findings
